+++ CHANGELOG.rst
@@ -1,6 +1,6 @@
 **0.1.21 - 6/16/25**
 
- - Update ...
+ - Update the "parallel" pipeline configuration key to "clones"
  - Update ParallelStep and EmbarrassinglyParallelStep classes to CloneableStep and AutoParallelStep, respectively
 
 **0.1.20 - 6/11/25**